Update
1. Changed enum structure on tiles to match the three overlaying meshes Floor, Wall, Overlay.
2. More renaming to account for three overlays.
3. Set up functions for Wall and Overlay mesh generation.  Currently not called anywhere.
4. Added logic to mesh data to determine which kind of mesh is being calculated.
5. Added functions to mesh data to calculate Wall and Overlay meshes and feed initial UVs.
6. Added functions to SpriteLoader to get Sprites for Walls and Overlays.  Renamed function for Floors.

Everything is working exactly as it did before adding logic for three meshes.  No additional meshes are actually being generated yet.  Still don't have a perfect system in place for idenifying all of the meshes that correspond to a particular tile but I think I have a solution.  It would be better if I could somehow store the mesh geometry on each mesh, it would simplify having to calculate geometry from the entire tile array and the relative position of the mesh and tile clicked every time an update is required.  The current method also requires that a world point is clicked in order to use it for geometry.  Still a million times better than the old RayCast method... but SO many calculations every click and can't be run in update/level load/etc.  This is a dealbreaker.  I need to find a solution that gets around this.