Update
1. Fully implemented full box selection.  You can now select in any direction and selection will grow and shrink accordingly.
2. Added function to MeshRefresh to add mesh gameobjects to the refresh list so that the refresh list doesn't need to be directly added to from public.
3. Bypassed (code is still present) the code remove duplicate meshes from the refresh list.  Instead, mesh game objects are only added to the list if the list does not already contain the same game object.  I'm not sure if Linq's .Distinct or standard .Contains is faster...  looks like Linq?  Both are pretty slow.

Next I want to implement a Box-Edges select.  If full box select is clearing an entire area out, Box-Edges select would build walls around a perimeter.  Only the tiles at the least and greatest x and y within the selection are selected.  Since this will be a second behavior, a method will be needed to select what behavior should be followed.  In order to do this, I will need to implement a primitive interface system (Designate, Build, etc)...  Fun times.